+++ flag/service/installation/installation.go
@@ -1,9 +1,10 @@
 package installation
 
 type Installation struct {
-	Customer string
-	Name     string
-	Pipeline string
-	Region   string
-	Registry string
+	Customer   string
+	InsecureCA string
+	Name       string
+	Pipeline   string
+	Region     string
+	Registry   string
 }