+++ pkg/remotewriteutils/util.go
@@ -29,7 +29,16 @@ func ToRemoteWrite(obj interface{}) (*pmov1alpha1.RemoteWrite, error) {
 }
 
 func FetchPrometheusList(ctx context.Context, r *ResourceWrapper, rw *pmov1alpha1.RemoteWrite) (*promv1.PrometheusList, error) {
-	selector, err := metav1.LabelSelectorAsSelector(&rw.Spec.ClusterSelector)
+	specSelector := &rw.Spec.ClusterSelector
+	// Adding an expression to ignore selecting prometheus-agent
+	ignoreAgentExp := metav1.LabelSelectorRequirement{
+		Key:      "app.kubernetes.io/name",
+		Operator: metav1.LabelSelectorOpNotIn,
+		Values:   []string{"prometheus-agent"},
+	}
+	specSelector.MatchExpressions = append(specSelector.MatchExpressions, ignoreAgentExp)
+	selector, err := metav1.LabelSelectorAsSelector(specSelector)
+
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}