+++ service/controller/clusterapi/controller.go
@@ -28,6 +28,7 @@ type ControllerConfig struct {
 	Bastions                []string
 	Customer                string
 	Installation            string
+	InsecureCA              bool
 	Pipeline                string
 	Provider                string
 	Region                  string