+++ service/controller/managementcluster/controller.go
@@ -29,6 +29,7 @@ type ControllerConfig struct {
 	Bastions                []string
 	Customer                string
 	Installation            string
+	InsecureCA              bool
 	Pipeline                string
 	Provider                string
 	Region                  string